requirements.txt
Flask==3.0.3
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
lxml==5.2.2
gunicorn==21.2.0
pytest==8.3.2
pytest-cov==5.0.0
